requirements.txt
Django>=5.2
djangorestframework>=3.14
django-cors-headers>=3.15

# Optional for token auth registration
djangorestframework-authtoken>=1.1
# dj-rest-auth and allauth are optional; code will work without them
# dj-rest-auth>=3.1
# django-allauth>=0.58
